Move Qwen position ID helpers to base
