+++ mlx_vlm/models/qwen3_vl_moe/qwen3_vl_moe.py
@@ -4,21 +4,11 @@ import mlx.core as mx
 import mlx.nn as nn
 import numpy as np
 
-from ..base import InputEmbeddingsFeatures
+from ..base import InputEmbeddingsFeatures, _prompt_position_ids
 from ..qwen3_vl import processing_qwen3_vl  # noqa: F401
 from .config import ModelConfig
 from .language import LanguageModel
 from .vision import VisionModel
-
-
-def _prompt_position_ids(position_ids: Optional[mx.array]) -> Optional[mx.array]:
-    if position_ids is None:
-        return None
-    if position_ids.ndim == 3 and position_ids.shape[0] == 3:
-        return position_ids.transpose(1, 2, 0)
-    if position_ids.ndim == 2:
-        return mx.broadcast_to(position_ids[:, :, None], (*position_ids.shape, 3))
-    return position_ids
 
 
 def masked_scatter(
